# registration-form-backend/tests/registration-form-backend.openapi.json
{"openapi":"3.0.1","info":{"title":"个人项目","description":"","version":"1.0.0"},"tags":[{"name":"hono报名表"},{"name":"hono报名表/账户"},{"name":"hono报名表/项目"},{"name":"hono报名表/报名"}],"paths":{"/auth/register":{"post":{"summary":"注册","deprecated":false,"description":"","tags":["hono报名表/账户"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"username":"testuser","email":"2116344865@qq.com","password":"password123","code":"729432"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/auth/logout":{"post":{"summary":"退出登录","deprecated":false,"description":"","tags":["hono报名表/账户"],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzU2MzU5MDIwNTQ0NjkiLCJyb2xlIjoidXNlciIsImlhdCI6MTczMzAyMTgxMX0.6WnFS86A14HRVEDqDaAJ3Mugo8srDcqFn4qNgwqboVo","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/auth/send-code":{"post":{"summary":"请求验证码","deprecated":false,"description":"","tags":["hono报名表/账户"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"email":"2116344865@qq.com"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/auth/login":{"post":{"summary":"登录","deprecated":false,"description":"","tags":["hono报名表/账户"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"email":"2116344865@qq.com","password":"password123"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/projects/delete/635827148353605":{"delete":{"summary":"管理员删除项目","deprecated":false,"description":"","tags":["hono报名表/项目"],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzU2MzU5MDIwNTQ0NjkiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3MzMwNjc1NjJ9.RakZ6bke8Iku7g3o3aLp3cwwYK7h7qU1YyDJNWYQrX8","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/projects/update/636093427957829":{"put":{"summary":"管理员修改项目","deprecated":false,"description":"","tags":["hono报名表/项目"],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzU2MzU5MDIwNTQ0NjkiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3MzMwNjc1NjJ9.RakZ6bke8Iku7g3o3aLp3cwwYK7h7qU1YyDJNWYQrX8","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"name":"更新后的项目名称","description":"新的描述","start_date":"2024-12-01","end_date":"2024-12-31","is_enabled":true}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/projects/create":{"post":{"summary":"管理员增加项目","deprecated":false,"description":"","tags":["hono报名表/项目"],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzYwOTMxOTY1MDkyNTMiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3MzMxMzMwMzh9.xb79N0Y-GbRXaVjfxG4U7c6ISIduT40W-J7uYfNiCEM","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"name":"项目名称1","description":"这是一个新项目","start_date":"2024-12-01","end_date":"2024-12-31","is_enabled":true}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/projects/query/635944967389253":{"get":{"summary":"获取单个项目","deprecated":false,"description":"","tags":["hono报名表/项目"],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzU2MzU5MDIwNTQ0NjkiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3MzMwNjc1NjJ9.RakZ6bke8Iku7g3o3aLp3cwwYK7h7qU1YyDJNWYQrX8","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/projects/query-list":{"get":{"summary":"获取项目列表","deprecated":false,"description":"","tags":["hono报名表/项目"],"parameters":[{"name":"page","in":"query","description":"","required":false,"example":"1","schema":{"type":"string"}},{"name":"pageSize","in":"query","description":"","required":false,"example":"5","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/registrations/update-status":{"put":{"summary":"管理员修改报名状态","deprecated":false,"description":"","tags":["hono报名表/报名"],"parameters":[{"name":"query_key","in":"query","description":"","required":false,"example":"d5f37217d693220e","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzYwOTMxOTY1MDkyNTMiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3MzMxMzMwMzh9.xb79N0Y-GbRXaVjfxG4U7c6ISIduT40W-J7uYfNiCEM","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"queryKey":"d5f37217d693220e","newStatus":1}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/registrations/query":{"get":{"summary":"按查询key或者联系方式查询","deprecated":false,"description":"","tags":["hono报名表/报名"],"parameters":[{"name":"query_key","in":"query","description":"","required":false,"example":"d5f37217d693220e","schema":{"type":"string"}},{"name":"contact_info","in":"query","description":"","required":false,"example":"bob@example.com","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/registrations/register":{"post":{"summary":"报名","deprecated":false,"description":"","tags":["hono报名表/报名"],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzYxNDgzMDU2ODY1OTciLCJyb2xlIjoidXNlciIsImlhdCI6MTczMzE0NjQ4Nn0.DJSpaTzqv07D0yGbhIs15fx3bxIZj-899glHRJ9pchI","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"projectId":"636093427957829","name":"Bob","contactInfo":"test@qq.com","isContactQueryAllowed":true}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/registrations/user-registrations":{"get":{"summary":"用户查询已报名","deprecated":false,"description":"","tags":["hono报名表/报名"],"parameters":[{"name":"project_id","in":"query","description":"","required":false,"example":"636158549233733","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzYxNDgzMDU2ODY1OTciLCJyb2xlIjoidXNlciIsImlhdCI6MTczMzE0Njc5NH0.bnr83sBZzEMhFBlC6UVwPR9Y5li2JaOVoyWDtWtnC5I","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/registrations/admin-registrations":{"get":{"summary":"管理员查询已报名","deprecated":false,"description":"","tags":["hono报名表/报名"],"parameters":[{"name":"project_id","in":"query","description":"","required":false,"example":"636158549233733","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzYwOTMxOTY1MDkyNTMiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3MzMxMzMwMzh9.xb79N0Y-GbRXaVjfxG4U7c6ISIduT40W-J7uYfNiCEM","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/registrations/cancel-registration":{"put":{"summary":"用户或临时用户取消报名","deprecated":false,"description":"","tags":["hono报名表/报名"],"parameters":[{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzYxNDgzMDU2ODY1OTciLCJyb2xlIjoidXNlciIsImlhdCI6MTczMzE1NDQ3MH0.BAvlD2YGUPzM8qJSgcQ8FGG8AiGi3ShlkdrDW-ilHhs","schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"queryKey":"798223bafdd8b7d9"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/registrations/verify-temporary-registration":{"put":{"summary":"临时用户修改报名状态","deprecated":false,"description":"","tags":["hono报名表/报名"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"email":"3280108463@qq.com","queryKey":"1ac55bfc1633c692","verificationCode":"530598"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/registrations/send-verification-code":{"post":{"summary":"临时用户修改报名状态验证码","deprecated":false,"description":"","tags":["hono报名表/报名"],"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{}},"example":{"email":"3280108463@qq.com","queryKey":"1ac55bfc1633c692"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}},"/registrations/export":{"get":{"summary":"管理员导出报名数据","deprecated":false,"description":"","tags":["hono报名表/报名"],"parameters":[{"name":"projectId","in":"query","description":"","required":false,"example":"636093427957829","schema":{"type":"string"}},{"name":"Authorization","in":"header","description":"","required":false,"example":"Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzYxNDgzMDU2ODY1OTciLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3MzMxNjA4MTh9.R42Wv4s5tbMEe8uyrMDEY90b5xF6Je4m6auIynKVv-M","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object","properties":{}}}},"headers":{}}},"security":[]}}},"components":{"schemas":{},"securitySchemes":{}},"servers":[]}